+++ zarrs/src/config.rs
@@ -22,7 +22,7 @@ use crate::array::{codec::CodecOptions, ArrayMetadataOptions};
 ///
 /// [`CodecOptions::validate_checksums()`] defaults to [`Config::validate_checksums()`].
 ///
-/// If validate checksums is enabled, checksum codecs (e.g. `crc32c`) will validate that encoded data matches stored checksums, otherwise validation is skipped.
+/// If validate checksums is enabled, checksum codecs (e.g. `crc32c`, `fletcher32`) will validate that encoded data matches stored checksums, otherwise validation is skipped.
 /// Note that regardless of this configuration option, checksum codecs may skip validation when partial decoding.
 ///
 /// ### Store Empty Chunks